requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.9.14
cssselect2==0.7.0
dj-database-url==1.0.0
Django==3.2.15
django-cors-headers==3.13.0
django-environ==0.9.0
django-filer==2.2.3
django-filter==22.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-polymorphic==3.1.0
djangorestframework==3.14.0
easy-thumbnails==2.8.3
gunicorn==20.1.0
lxml==4.9.1
Pillow==9.2.0
pip==22.1.2
psycopg2==2.8.6
pycodestyle==2.9.1
pytz==2022.1
reportlab==3.6.11
setuptools==63.4.1
sqlparse==0.4.1
svglib==1.4.1
tinycss2==1.1.1
toml==0.10.2
Unidecode==1.1.2
webencodings==0.5.1
wheel==0.37.1
whitenoise==6.2.0
